+++ umh-core/pkg/fsm/benthos_monitor/actions.go
@@ -31,7 +31,7 @@ func (b *BenthosMonitorInstance) CreateInstance(ctx context.Context, filesystemS
 	b.baseFSMInstance.GetLogger().Debugf("Starting Action: Adding Benthos Monitor service %s to S6 manager ...", b.baseFSMInstance.GetID())
 
 	b.baseFSMInstance.GetLogger().Debugf("Adding Benthos Monitor service %s to S6 manager with port %d", b.baseFSMInstance.GetID(), b.config.MetricsPort)
-	err := b.monitorService.AddBenthosMonitorToS6Manager(ctx, uint16(b.config.MetricsPort))
+	err := b.monitorService.AddBenthosMonitorToS6Manager(ctx, b.config.MetricsPort)
 	if err != nil {
 		if err == benthos_monitor_service.ErrServiceAlreadyExists {
 			b.baseFSMInstance.GetLogger().Debugf("Benthos Monitor service %s already exists in S6 manager", b.baseFSMInstance.GetID())
